+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Adding validation for worker instance types in `AWSMachinedeployment` CR.
 - Prevent max number of nodes in `AWSMachinedeployment` CR to be 0 or smaller than min number of nodes.
 - Prevent upgrades (changing release version label on `Cluster` CR) if the Cluster has not transitioned yet.